+++ src_3_0_0/core/layers/js/layer2dRasterWmsTime.js
@@ -206,6 +206,23 @@ Layer2dRasterWmsTimeLayer.prototype.incrementIsSmaller = function (step, increme
 };
 
 /**
+ * Creates the capabilities url.
+ * @param {String} wmsTimeUrl The url of wms time.
+ * @param {String} version The version of wms time.
+ * @param {String} layers The layers of wms time.
+ * @returns {String} the created url
+ */
+Layer2dRasterWmsTimeLayer.prototype.createCapabilitiesUrl = function (wmsTimeUrl, version, layers) {
+    const url = new URL(wmsTimeUrl);
+
+    url.searchParams.set("service", "WMS");
+    url.searchParams.set("version", version);
+    url.searchParams.set("layers", layers);
+    url.searchParams.set("request", "GetCapabilities");
+    return url;
+};
+
+/**
  * Requests the GetCapabilities document and parses the result.
  * @param {String} url The url of wms time.
  * @param {String} version The version of wms time.
@@ -213,7 +230,7 @@ Layer2dRasterWmsTimeLayer.prototype.incrementIsSmaller = function (step, increme
  * @returns {Promise} A promise which will resolve the parsed GetCapabilities object.
  */
 Layer2dRasterWmsTimeLayer.prototype.requestCapabilities = function (url, version, layers) {
-    return axios.get(encodeURI(`${url}?service=WMS&version=${version}&layers=${layers}&request=GetCapabilities`))
+    return axios.get(this.createCapabilitiesUrl(url, version, layers))
         .then(response => handleAxiosResponse(response, "WMS, createLayerSource, requestCapabilities"));
 };